examples/servers/mcpdoc/requirements.txt
aws-cdk-lib==2.204.0
aws-cdk.aws-lambda-python-alpha==2.204.0a0
constructs>=10.0.0,<11.0.0
cdk-nag==2.36.33
